# swagger.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "{scheme}://{host}/api/v1",
      "variables": {
        "scheme": {
          "enum": [
            "https",
            "http"
          ],
          "default": "https"
        },
        "host": {
          "default": "localhost:3000"
        }
      }
    },
    {
      "url": "/api/v1"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "mycelia-api",
    "description": "",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "paths": {
...
    "/auth/signup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "signup",
          "signUp"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "login",
          "signIn"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "forgot_password",
          "forgotPassword"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "reset_password",
          "resetPassword"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/verify-email": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "verify_email",
          "verifyEmail"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
...
    "/auth/delete": {
      "delete": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "auth",
          "$delete",
          "deleteAccount"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/update-email": {
      "patch": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "auth",
          "update_email",
          "updateEmail"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/update-username": {
      "patch": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "auth",
          "update_username",
          "updateUsername"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/update-password": {
      "patch": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "auth",
          "update_password",
          "updatePassword"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/profile": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "auth",
          "profile",
          "getProfile"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/verify-totp": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "verify_totp",
          "verifyTotp"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/mfa/totp/verify": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "mfa",
          "totp",
          "verify",
          "verifyTotpNew"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/mfa/webauthn/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "auth",
          "mfa",
          "webauthn",
          "register",
          "startWebAuthnRegistration"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/mfa/webauthn/callback": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "auth",
          "mfa",
          "webauthn",
          "callback",
          "finishWebAuthnRegistration"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/mfa/webauthn/verify": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "x-samchon-accessor": [
          "auth",
          "mfa",
          "webauthn",
          "verify",
          "verifyWebAuthnAssertion"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/mfa/totp/enroll": {
      "patch": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "auth",
          "mfa",
          "totp",
          "enroll",
          "enrollTotp"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
    "/auth/mfa/totp/disable": {
      "patch": {
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "auth",
          "mfa",
          "totp",
          "disable",
          "disableTotp"
        ],
        "x-samchon-controller": "AuthController"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}